+++ src/main/java/uk/ac/ebi/jmzml/model/mzml/BinaryDataArrayList.java
@@ -3,7 +3,6 @@ package uk.ac.ebi.jmzml.model.mzml;
 
 import javax.xml.bind.annotation.*;
 import java.io.Serializable;
-import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -44,7 +43,7 @@ public class BinaryDataArrayList
     protected List<BinaryDataArray> binaryDataArray;
     @XmlAttribute(required = true)
     @XmlSchemaType(name = "nonNegativeInteger")
-    protected BigInteger count;
+    protected Integer count;
 
     /**
      * Gets the value of the binaryDataArray property.
@@ -80,10 +79,10 @@ public class BinaryDataArrayList
      * 
      * @return
      *     possible object is
-     *     {@link BigInteger }
+     *     {@link Integer }
      *     
      */
-    public BigInteger getCount() {
+    public Integer getCount() {
         return count;
     }
 
@@ -92,10 +91,10 @@ public class BinaryDataArrayList
      * 
      * @param value
      *     allowed object is
-     *     {@link BigInteger }
+     *     {@link Integer }
      *     
      */
-    public void setCount(BigInteger value) {
+    public void setCount(Integer value) {
         this.count = value;
     }
