+++ src/main/java/uk/ac/ebi/jmzml/model/mzml/IndexList.java
@@ -1,16 +1,10 @@
 
 package uk.ac.ebi.jmzml.model.mzml;
 
+import javax.xml.bind.annotation.*;
 import java.io.Serializable;
-import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlAttribute;
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlSchemaType;
-import javax.xml.bind.annotation.XmlType;
 
 
 /**
@@ -47,7 +41,7 @@ public class IndexList
     protected List<Index> index;
     @XmlAttribute(required = true)
     @XmlSchemaType(name = "nonNegativeInteger")
-    protected BigInteger count;
+    protected Integer count;
 
     /**
      * Gets the value of the index property.
@@ -83,10 +77,10 @@ public class IndexList
      * 
      * @return
      *     possible object is
-     *     {@link BigInteger }
+     *     {@link Integer }
      *     
      */
-    public BigInteger getCount() {
+    public Integer getCount() {
         return count;
     }
 
@@ -95,10 +89,10 @@ public class IndexList
      * 
      * @param value
      *     allowed object is
-     *     {@link BigInteger }
+     *     {@link Integer }
      *     
      */
-    public void setCount(BigInteger value) {
+    public void setCount(Integer value) {
         this.count = value;
     }
